requirements.txt
discord.py==2.3.1
tomlkit==0.11.8
cchardet==2.1.7